// frontend/src/api/generated/models/allocation-order-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AllocationOrderSummary } from './allocation-order-summary';

/**
 * 
 * @export
 * @interface AllocationOrderPageResponse
 */
export interface AllocationOrderPageResponse {
    /**
     * 
     * @type {Array<AllocationOrderSummary>}
     * @memberof AllocationOrderPageResponse
     */
    'content'?: Array<AllocationOrderSummary>;
    /**
     * 
     * @type {number}
     * @memberof AllocationOrderPageResponse
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof AllocationOrderPageResponse
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof AllocationOrderPageResponse
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof AllocationOrderPageResponse
     */
    'totalPages'?: number;
}