// frontend/src/api/generated/models/allocation-execution-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AllocatedSlipDetail } from './allocated-slip-detail';
// May contain unused imports in some cases
// @ts-ignore
import type { UnallocatedLineDetail } from './unallocated-line-detail';
// May contain unused imports in some cases
// @ts-ignore
import type { UnpackInstructionItem } from './unpack-instruction-item';

/**
 * 
 * @export
 * @interface AllocationExecutionResult
 */
export interface AllocationExecutionResult {
    /**
     * 引当成功件数
     * @type {number}
     * @memberof AllocationExecutionResult
     */
    'allocatedCount'?: number;
    /**
     * 
     * @type {Array<AllocatedSlipDetail>}
     * @memberof AllocationExecutionResult
     */
    'allocatedSlips'?: Array<AllocatedSlipDetail>;
    /**
     * 
     * @type {Array<UnpackInstructionItem>}
     * @memberof AllocationExecutionResult
     */
    'unpackInstructions'?: Array<UnpackInstructionItem>;
    /**
     * 
     * @type {Array<UnallocatedLineDetail>}
     * @memberof AllocationExecutionResult
     */
    'unallocatedLines'?: Array<UnallocatedLineDetail>;
}