// frontend/src/api/generated/models/allocated-order-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AllocatedOrderSummary
 */
export interface AllocatedOrderSummary {
    /**
     * 出荷伝票ID
     * @type {number}
     * @memberof AllocatedOrderSummary
     */
    'id'?: number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof AllocatedOrderSummary
     */
    'slipNumber'?: string;
    /**
     * 出荷予定日
     * @type {string}
     * @memberof AllocatedOrderSummary
     */
    'plannedDate'?: string;
    /**
     * 出荷先取引先名
     * @type {string}
     * @memberof AllocatedOrderSummary
     */
    'partnerName'?: string;
    /**
     * 全明細件数
     * @type {number}
     * @memberof AllocatedOrderSummary
     */
    'lineCount'?: number;
    /**
     * 引当済み明細件数
     * @type {number}
     * @memberof AllocatedOrderSummary
     */
    'allocatedLineCount'?: number;
    /**
     * ステータス
     * @type {string}
     * @memberof AllocatedOrderSummary
     */
    'status'?: AllocatedOrderSummaryStatusEnum;
}

export const AllocatedOrderSummaryStatusEnum = {
    Allocated: 'ALLOCATED',
    PartialAllocated: 'PARTIAL_ALLOCATED'
} as const;

export type AllocatedOrderSummaryStatusEnum = typeof AllocatedOrderSummaryStatusEnum[keyof typeof AllocatedOrderSummaryStatusEnum];