// frontend/src/api/generated/models/allocated-order-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AllocatedOrderSummary } from './allocated-order-summary';

/**
 * 
 * @export
 * @interface AllocatedOrderPageResponse
 */
export interface AllocatedOrderPageResponse {
    /**
     * 
     * @type {Array<AllocatedOrderSummary>}
     * @memberof AllocatedOrderPageResponse
     */
    'content'?: Array<AllocatedOrderSummary>;
    /**
     * 
     * @type {number}
     * @memberof AllocatedOrderPageResponse
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof AllocatedOrderPageResponse
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof AllocatedOrderPageResponse
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof AllocatedOrderPageResponse
     */
    'totalPages'?: number;
}