// frontend/src/api/generated/models/unpack-instruction-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UnpackInstructionItem
 */
export interface UnpackInstructionItem {
    /**
     * ばらし指示ID
     * @type {number}
     * @memberof UnpackInstructionItem
     */
    'id'?: number;
    /**
     * 商品コード
     * @type {string}
     * @memberof UnpackInstructionItem
     */
    'productCode'?: string;
    /**
     * 商品名
     * @type {string}
     * @memberof UnpackInstructionItem
     */
    'productName'?: string;
    /**
     * 元荷姿（CASE / BALL）
     * @type {string}
     * @memberof UnpackInstructionItem
     */
    'fromUnitType'?: string;
    /**
     * 先荷姿（BALL / PIECE）
     * @type {string}
     * @memberof UnpackInstructionItem
     */
    'toUnitType'?: string;
    /**
     * ばらし数量
     * @type {number}
     * @memberof UnpackInstructionItem
     */
    'quantity'?: number;
    /**
     * ステータス
     * @type {string}
     * @memberof UnpackInstructionItem
     */
    'status'?: UnpackInstructionItemStatusEnum;
}

export const UnpackInstructionItemStatusEnum = {
    Instructed: 'INSTRUCTED'
} as const;

export type UnpackInstructionItemStatusEnum = typeof UnpackInstructionItemStatusEnum[keyof typeof UnpackInstructionItemStatusEnum];